refactor: extract Color to its own file
