+++ types/IwFormColor.d.ts
@@ -1,0 +1,6 @@
+type IwFormColorHexString = `#${string}`
+
+type IwFormColorRgbArray = [number, number, number] | [number, number, number, number]
+
+type IwFormColorRgbString = `${'RGB' | 'rgb'}(${number},${number},${number})`
+    | `${'RGBA' | 'rgba'}(${number},${number},${number},${number})`
